requirements.txt
attrs==19.1.0
baidu-aip==2.2.17.0
billiard==3.6.0.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.4
dill==0.2.9
editdistance==0.5.3
flasgger==0.9.3
Flask==1.1.1
Flask-Cors==3.0.8
flask-peewee==3.0.3
Flask-SQLAlchemy==2.4.0
future==0.16.0
gevent==1.4.0
google-auth==1.6.3
greenlet==0.4.15
grpcio==1.23.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
mistune==0.8.4
peewee==3.11.2
ply==3.8
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyMySQL==0.9.3
pyrsistent==0.15.4
python-gnupg==0.4.3
pytz==2019.2
PyYAML==5.1.2
pyjwt==1.5.3
redis==3.3.6
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
wtf-peewee==3.0.0
WTForms==2.2.1
xmltodict==0.12.0
bcrypt==3.1.7

